Split ARRIS Surfboard HTTP and HNAP providers
