+++ src/NetworkOptimizer.Web/Components/Pages/Settings.razor
@@ -1001,6 +1001,7 @@
                             <select class="form-control" @bind="_editingCm.Provider">
                                 <option value="netgear">Netgear CM (HTTP)</option>
                                 <option value="arris-surfboard">ARRIS Surfboard (HTTP)</option>
+                                <option value="arris-surfboard-hnap">ARRIS Surfboard (HNAP)</option>
                                 <option value="motorola-hnap">Motorola MB8611, MB8600 (HNAP)</option>
                             </select>
                         </div>
@@ -4181,6 +4182,7 @@
     {
         "netgear" => "Netgear CM (HTTP)",
         "arris-surfboard" => "ARRIS Surfboard (HTTP)",
+        "arris-surfboard-hnap" => "ARRIS Surfboard (HNAP)",
         "motorola-hnap" => "Motorola (HNAP)",
         _ => provider,
     };
@@ -4188,6 +4190,7 @@
     private static string GetCmStatusPathPlaceholder(string provider) => provider switch
     {
         "arris-surfboard" => "/cgi-bin/status",
+        "arris-surfboard-hnap" => "N/A (uses /HNAP1/)",
         "motorola-hnap" => "N/A (uses /HNAP1/)",
         _ => "/DocsisStatus.htm",
     };
